requirements.txt

streamlit==1.14.0
requests==2.25.1
PyYAML==5.4.1
prance==0.19.0
openapi-spec-validator==0.3.1
